Add movies.js file
